+++ src/include/sll/api/_generated.h
@@ -135,6 +135,28 @@
 
 
 
+#define __SLL_API_TYPE_sll_api_file_flush void
+#define __SLL_API_ARGS_sll_api_file_flush sll_handle_data_t* a
+/**
+ * \flags check_output func optimizable
+ * \name sll_api_file_flush
+ * \group file-api
+ * \desc Docs!
+ * \arg sll_handle_data_t* a -> 
+ */
+/**
+ * \flags check_output func optimizable
+ * \name sll_api_file_flush_raw
+ * \group raw-api
+ * \subgroup raw-api-file-api
+ * \desc Wrapper function for :sll_api_file_flush:
+ * \arg sll_object_t*const* al -> Arguments
+ * \arg sll_arg_count_t all -> Argument count
+ * \ret sll_object_t* -> The return value of the function
+ */
+
+
+
 #define __SLL_API_TYPE_sll_api_file_open void
 #define __SLL_API_ARGS_sll_api_file_open sll_string_t* a,sll_string_t* b,sll_handle_data_t* out
 /**