+++ src/sll/api/_generated.c
@@ -172,6 +172,30 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_file_close_raw(sll_objec
 
 
 
+__SLL_API_TYPE_sll_api_file_flush sll_api_file_flush(__SLL_API_ARGS_sll_api_file_flush);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_file_flush_raw(sll_object_t*const* al,sll_arg_count_t all){
+	sll_object_t* a=NULL;
+	if (all>0){
+		a=*(al+0);
+		if ((SLL_OBJECT_GET_TYPE(a)==SLL_OBJECT_TYPE_HANDLE&&a->dt.h.t!=SLL_HANDLE_UNKNOWN_TYPE)){
+			SLL_ACQUIRE(a);
+		}
+		else{
+			a=sll_operator_cast(a,sll_static_int[SLL_OBJECT_TYPE_HANDLE]);
+		}
+	}
+	else{
+		a=SLL_CREATE();
+		a->t=SLL_OBJECT_TYPE_HANDLE;
+		SLL_INIT_HANDLE_DATA(&(a->dt.h));
+	}
+	sll_api_file_flush(&(a->dt.h));
+	SLL_RELEASE(a);
+	return SLL_ACQUIRE_STATIC_INT(0);
+}
+
+
+
 __SLL_API_TYPE_sll_api_file_open sll_api_file_open(__SLL_API_ARGS_sll_api_file_open);
 __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_file_open_raw(sll_object_t*const* al,sll_arg_count_t all){
 	sll_object_t* a=NULL;
@@ -1668,6 +1692,7 @@ static const internal_function_t _ifunc_data_ptr[]={
 	{"sll:array_replace",sll_api_array_replace_raw,SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL},
 	{"sll:array_reverse",sll_api_array_reverse_raw,SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL},
 	{"sll:file_close",sll_api_file_close_raw,SLL_INTERNAL_FUNCTION_FLAG_REQUIRED},
+	{"sll:file_flush",sll_api_file_flush_raw,SLL_INTERNAL_FUNCTION_FLAG_REQUIRED},
 	{"sll:file_open",sll_api_file_open_raw,0},
 	{"sll:file_read",sll_api_file_read_raw,SLL_INTERNAL_FUNCTION_FLAG_REQUIRED},
 	{"sll:file_std_handle",sll_api_file_std_handle_raw,0},
@@ -1724,5 +1749,5 @@ static const internal_function_t _ifunc_data_ptr[]={
 
 
 
-const sll_function_index_t _ifunc_size=57;
+const sll_function_index_t _ifunc_size=58;
 const internal_function_t* _ifunc_data=(const internal_function_t*)(&_ifunc_data_ptr);